Address review: fold permalink helper into backend.go, fix stale comments
